app/requirements.txt
alembic==1.0.10
bleach==3.1.0
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.7.0
Flask-PageDown==0.2.2
Flask-QRcode==3.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-SSLify==0.1.5
Flask-thumbnails==1.0.6
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
Markdown==3.1.1
MarkupSafe==1.1.1
mollie-api-python==2.1.3
mysqlclient==1.4.2.post1
Pillow==6.0.0
python-dateutil==2.8.0
python-editor==1.0.4
qrcode==6.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
unicodecsv==0.14.1
urllib3==1.25.3
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.15.4
WTForms==2.2.1
